Create hotelcost.py
There are N hotels along the beautiful Adriatic coast. Each hotel has its value in Euros.
Sroljo has won M Euros on the lottery. Now he wants to buy a sequence of consecutive hotels,
such that the sum of the values of these consecutive hotels is as great as possible 
but not greater than M.
calculate this greatest possible total value.